PlayerCardコンポーネントのリファクタリングとPlayerStatusCardコンポーネントの削除
